형태에 따른 기능 분류 추가
